+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+### Added
+
+- Import Catalogue filter now allows selecting multiple filters at once.
+
 ## [7.0.8] - 2022-03-08
 
 ### Fixed